% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfBetas.R
\name{dfBetas}
\alias{dfBetas}
\title{Return DFBetas or DFBeta values}
\usage{
dfBetas(obj, key.variable = NULL, sort = FALSE, standardize = TRUE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{key.variable}{Used if lm() data object is of tibble class. Name of the unique key variable (identifier variable). If data object is of data.frame class, row.names will be used instead.}

\item{sort}{Logical: Should the returned dataframe be sorted by the number of tagged coefficients?}

\item{standardize}{Logicial: Should standardized DFBeta values (DFBetas) be returned in the dataframe?}
}
\description{
Simple function to produce a listing of all DFBetas (or DFBeta) values for the coefficients in a lm() object. Observations to investigate (tagged as Investigate) are row sums of absolute coefficients greater than 2 / sqrt(N).
}
\examples{

mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
dfBetas(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{levPlot}}
\code{\link{cooksPlot}}
\code{\link{threeOuts}}
}

