% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Create a Histogram}
\usage{
histogram(x, breaks = NULL, table = FALSE, right = FALSE,
  xlab = deparse(substitute(x)), title = NULL, ...)
}
\arguments{
\item{x}{Variable to be plotted.}

\item{breaks}{one of:
\itemize{
  \item a vector giving the breakpoints between histogram cells,
  \item a function to compute the vector of breakpoints,
  \item a single number giving the number of cells for the histogram,
  \item a character string naming an algorithm to compute the number of cells,
  \item a function to compute the number of cells.
  }}

\item{table}{Logical: Should a frequency distribution table be produced?}

\item{right}{Logical; if FALSE, the histogram cells are left-closed (right open) intervals.}

\item{xlab}{Label of the x-axis. Default is the object name of x above.}

\item{title}{Title (optional) of the chart.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Functional wrapper to hist() in graphics. Adds functionality (frequency table) and standard cut options (right=FALSE) to hist().
}
\examples{
histogram(mtcars$hp)
histogram(mtcars$hp, breaks=seq(50, 350, by=25))
histogram(mtcars$hp, breaks=seq(50, 350, by=25), table=TRUE)

}

