% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertChunkBreak.R
\name{insertChunkBreak}
\alias{insertChunkBreak}
\title{Insert break in chunk.}
\usage{
insertChunkBreak()
}
\description{
Call this function as an addin to insert an *R code chunk break* at the cursor position.
}

