% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertCodefont.R
\name{insertCodefont}
\alias{insertCodefont}
\title{Insert break in chunk.}
\usage{
insertCodefont()
}
\description{
Call this function as an addin to insert an *R code chunk break* at the cursor position.
}
