% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levPlot.R
\name{levPlot}
\alias{levPlot}
\title{Create a Leverage (hat values) Plot}
\usage{
levPlot(obj, ylim = NULL, key.variable = NULL, print.obs = FALSE,
  print.plot = TRUE, sort.obs = FALSE, save.cutoff = FALSE,
  cut.level = 3)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{ylim}{Optional ylim for the plot.}

\item{key.variable}{Required if lm() data object is of tibble class. Name of the unique key variable (identifier variable). If data object is of data.frame class, this is optional and row.names will be used instead.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}

\item{sort.obs}{Logical: Should observations (if print.obs=TRUE) be sorted?}

\item{save.cutoff}{Logical: Should the used cutoff of the hat value be saved. The standard cut-offs (graphed) at 2 and 3 times the average hat value ($(k+1) / n$).}

\item{cut.level}{Level to be used in *printing* the list of Leverage outliers. Defaults to 3 times the average hat value ($(k+1) / n$).}
}
\description{
Simple function to produce a plot of the hat values (leverage) after a model run in lm().
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
leveragePlot(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{cooksPlot}}
\code{\link{threeOuts}}
}
