% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmSingleR2.R
\name{lmSingleR2}
\alias{lmSingleR2}
\title{Return an R2 value for lm model variable}
\usage{
lmSingleR2(obj, test.var)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{test.var}{Required. Variable name from lm() object.}
}
\description{
Simple function to produce a single \eqn{R^2} value for a specified variable (including factors and interaction terms) from a lm() object. Uses anova() to generate type 1 Sums of Squares (sequential) for variable.
}
\examples{

mtcars$gear <- factor(mtcars$gear)
mod <- lm(hp ~ disp + gear, data=mtcars)
summary(mod)
lmSingleR2(mod, "gear")

}
\seealso{
\code{\link{pCorr}}
}
