% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDummies.R
\name{makeDummies}
\alias{makeDummies}
\title{Create Dummy Codes}
\usage{
makeDummies(x)
}
\arguments{
\item{x}{Vector to be Dummy Coded.}
}
\description{
Simple function to create Dummy Codes from a single variable. All "k" Dummy Variables will be created.
}
\examples{

makeDummies(mtcars$cyl)
mtcars <- data.frame(mtcars, makeDummies(mtcars$cyl))

}
\seealso{
\code{\link{makeEffects}}
}
