% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEffects.R
\name{makeEffects}
\alias{makeEffects}
\title{Create Effect Codes}
\usage{
makeEffects(x, ref)
}
\arguments{
\item{x}{Vector to be Dummy Coded.}

\item{ref}{Name of the base reference category.}
}
\description{
Simple function to create Effect Codes from a single variable, given some base reference category. Only "k-1" Effect Variables will be created - excluding the base reference group.
}
\examples{

makeEffects(mtcars$cyl, "4")
mtcars <- data.frame(mtcars, makeEffects(mtcars$cyl, "4"))

}
\seealso{
\code{\link{makeDummies}}
}
