% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleAnova.R
\name{simpleAnova}
\alias{simpleAnova}
\title{Simple Anova Table}
\usage{
simpleAnova(obj, ...)
}
\arguments{
\item{obj}{Model object from an \code{lm()} fitted equation.}

\item{...}{Additional \code{anova()} options if needed.}
}
\description{
Simple function to produce an ANOVA table (Model and Error) after a model run in \code{lm()}. This replicates the F-test in the \code{summary.lm()} function.
}
\section{References}{


Found via the interwebs \href{https://stats.stackexchange.com/questions/145790/anova-table-for-model-in-r}{here}.
}

\examples{

mod <- lm(hp ~ disp, data=mtcars)
simpleAnova(mod)
summary(mod)

}
\seealso{
\code{\link{lm}}
}
