% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlane.R
\name{simplePlane}
\alias{simplePlane}
\title{Create a simple plane}
\usage{
simplePlane(obj)
}
\arguments{
\item{obj}{Model object from lm().}
}
\description{
Useful in describing a trivariate model. Used only in slides...
}
\keyword{internal}
