% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standResidPlot.R
\name{standResidPlot}
\alias{standResidPlot}
\title{Create a Standardized Residual Plot}
\usage{
standResidPlot(obj, id = FALSE, print.obs = FALSE, print.plot = TRUE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{id}{Logical: Should the identy() function be used to describe selected observations? Will automatically invoke print=TRUE for selected observations.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}
}
\description{
Produce a Standardized Residual plot. NOT recommended. Use Studentized Deleted instead
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
standResidPlot(mod)

}
\seealso{
\code{\link{studResidPlot}}
}
\keyword{internal}

