% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studResidPlot.R
\name{studResidPlot}
\alias{studResidPlot}
\title{Create a Studentized Deleted Residual Plot}
\usage{
studResidPlot(obj, id = FALSE, print.obs = FALSE, print.plot = TRUE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{id}{Logical: Should the identy() function be used to describe selected observations? Will automatically invoke print=TRUE for selected observations.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}
}
\description{
Produce a studentized deleted residual plot.
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
studResidPlot(mod)
}
\seealso{
\code{\link{levPlot}}
\code{\link{cooksPlot}}
\code{\link{threeOuts}}
}

