\name{tagObs}
\alias{tagObs}
\docType{package}
\title{Tag Observations}
\description{A minimal function to mark observations use in a prior run lm() model. Observations will be marked with a new variable in the original data used in the lm() function.}
\usage{
  tagObs(model, tag=1, name=NULL)
}
\arguments{
  \item{model}{ a model object generated by the lm() function. }
  \item{tag}{ label with which to mark the used observations. Defualt is an indicator of 1. }
  \item{name}{ option name to assign to the suffix of the newly created varaible. Defualt is the model object name. }
  \item{...}{ further arguments passed to or from other methods. }
}
\keyword{ package }
\note{simpleScatter produces a ggplot object that may be used in conjunction with ggplot followup commands.}
\examples{
  mod <- lm(hp ~ disp + wt, data=mtcars)
  tagObs(model = mod)
}
