% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeOuts.R
\name{threeOuts}
\alias{threeOuts}
\title{Combine All Three Common Outlier Types}
\usage{
threeOuts(obj, key.variable = NULL, print.plot = FALSE)
}
\arguments{
\item{obj}{Model object from an lm() fiitted equation.}

\item{key.variable}{Used if lm() data object is of tibble class. Name of the unique key variable (identifier variable). If data object is of data.frame class, row.names will be used instead.}

\item{print.plots}{Logical: Should plots for all three outlier tests be shown?}
}
\description{
Simple function to produce a plots and a final data frame from studentized deleted residuals, hat values (leverage), and Cook's distance, after a model run in lm(). Resulting dataframe will be sorted by the inThree variable (observation is considered an outlier on all three plots) and then the Cooks_D variable.
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
threeOuts(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{levPlot}}
\code{\link{cooksPlot}}
}
