% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSubtitle.R
\name{addSubtitle}
\alias{addSubtitle}
\title{addSubtitle}
\usage{
addSubtitle(title, subtitle)
}
\arguments{
\item{title}{Title to be added to the existing ggplot object.}

\item{subtitle}{Subtitle to be added to existing ggplot object.
... Further agruements passed to or from other methods.}
}
\description{
A simple function to add a title and subtitle to an existing ggplot object.
}
\examples{
m <- mtcars
s <- simpleScatter(m, disp)
s + addSubtitle("Horsepower from Displacement", "From mtcars dataset")
}
\seealso{
\code{\link{simpleScatter}}
}

