% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooksPlot.R
\name{cooksPlot}
\alias{cooksPlot}
\title{Create a Cook's Distance Plot}
\usage{
cooksPlot(
  obj,
  ylim = NULL,
  key.variable = NULL,
  print.obs = FALSE,
  print.plot = TRUE,
  sort.obs = TRUE,
  all.obs = FALSE,
  save.cutoff = FALSE
)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{ylim}{Optional ylim for the plot.}

\item{key.variable}{Required if lm() data object is of tibble class. Name of the unique key variable (identifier variable). If data object is of data.frame class, this is optional and row.names will be used instead.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}

\item{sort.obs}{Logical: Should observations (if print.obs=TRUE) be sorted?}

\item{all.obs}{Logical: Should \strong{all} observations be printed?}

\item{save.cutoff}{Logical: Should the Cook's Distance cutoff be saved to the Global Environment?}
}
\description{
Simple function to produce a plot of the Cook's Distance for each observation after a model run in lm().
}
\examples{

mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
cooksPlot(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{levPlot}}
\code{\link{threeOuts}}
}
