% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooksPlot.R
\name{cooksPlot}
\alias{cooksPlot}
\title{cooksPlot}
\usage{
cooksPlot(obj, ylim = NULL, id = FALSE, print.obs = FALSE,
  print.plot = TRUE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{ylim}{Optional ylim for the plot.}

\item{id}{Logical: Should the identy() function be used to describe selected observations? Will automatically invoke print=TRUE for selected observations.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}
}
\description{
Simple function to produce a plot of the Cook's Distance for each observation after a model run in lm().
}
\examples{

mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
cooksPlot(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{leveragePlot}}
\code{\link{threeOuts}}
}

