% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveragePlot.R
\name{leveragePlot}
\alias{leveragePlot}
\title{Create a Leverage (hat values) Plot}
\usage{
leveragePlot(obj, ylim = NULL, id = FALSE, print.obs = FALSE,
  print.plot = TRUE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{ylim}{Optional ylim for the plot.}

\item{id}{Logical: Should the identy() function be used to describe selected observations? Will automatically invoke print=TRUE for selected observations.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}
}
\description{
Simple function to produce a plot of the hat values (leverage) after a model run in lm().
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
leveragePlot(mod)

}
\seealso{
\code{\link{studResidPlot}}
\code{\link{cooksPlot}}
\code{\link{threeOuts}}
}

