% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmBeta.R
\name{lmBeta}
\alias{lmBeta}
\title{Standardized Betas}
\usage{
lmBeta(obj)
}
\arguments{
\item{obj}{Model object from an \code{lm()} fitted equation.}
}
\description{
Simple function to produce standardized beta estimates after a model run in \code{lm()}.
}
\examples{
mod <- lm(hp ~ disp + wt, data=mtcars)
summary(mod)
lmBeta(mod)

}
