% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmDecomp.R
\name{lmDecomp}
\alias{lmDecomp}
\title{Decompose lm() Interactions}
\usage{
lmDecomp(obj, interest, moderator, mod.type = 1, int.values = NULL,
  mod.values = NULL, int.range = NULL, mod.range = NULL,
  alpha.level = 0.05, show.points = FALSE, print.sslopes = TRUE,
  print.ros = FALSE)
}
\arguments{
\item{obj}{Model object containing either a Quantitative by Quantitative interaction or a Quantitative by Categorical interaction from lm().}

\item{interest}{Variable of interest in the model. Must use quotes.}

\item{moderator}{Variable designated as the moderator in the model. Must use quotes.}

\item{mod.type}{Defines the TYPE of interaction. Must be an integer of 1 or 2. 1 = Quantitative by Quantitative. 2 = Quantitative by Categorical. Defaults to 1.}

\item{int.values}{DEPRECIATED. Set to NULL.}

\item{mod.values}{Defines the values of the moderator to be used when calculating simple slopes. Only used with mod.type=1. If mod.type=2, mod.values is set to *dummy* coding (0,1), and comparison will be between indicator level and reference level in the model.}

\item{int.range}{Optional. Defines the range of the variable of interest to be graphed. Only used when mod.type=2.}

\item{mod.range}{Optional. Defines the range of the moderator to be graphed.}

\item{alpha.level}{Designates the alpha for the Region-of-Significnace. Defaults to 0.5.}

\item{show.points}{Logical. Should the original data points be shown on the produced graph? Defaults to FALSE.}

\item{print.sslopes}{Logical. Should the Simple Slopes graph be created? Defaults to TRUE.}

\item{print.ros}{Logical. Should the Region-of-Significance values and graph be created? Defaults to FALSE.}
}
\description{
This function takes the work of Bauer and Curan (2005) and applies it to a linear model object produced by lm(). Currently, the function handles two types of interactions in the lm() model: (1) Quantitative by Quantitative, which produces the Aiken and West (1991) Pick-a-Point approach simple slopes as well as Johnson-Neyman Regions-of-Significance; and (2) Quantitative by Categorical, which produces the simple slopes for the two levels of the dummy factor level as well as a quasi Region-of-Significance marking values of the Quantitative variable at which the difference of the factor levels change significance.
}
\details{
All graphs produced are exported to the Global Environment as ggplot object which can be further edited.
}
\examples{
mod <- lm(mpg ~ wt * am, data=mtcars)
summary(mod)
lmDecomp(mod, "wt", "am", mod.type = 2, mod.values=c(0,1))

}
\references{
Bauer, D.J. & Curran, P.J., (2005), Probing Interactions in Fixed and Multilevel Regression: Inferential and Graphical Techniques, Multivariate Behavioral Research, 40(3), 373–400.
}
