% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelData.R
\name{modelData}
\alias{modelData}
\title{Create a \emph{FULL} dataset based on a model object}
\usage{
modelData(obj)
}
\arguments{
\item{obj}{Model object from an \code{lm()} fitted equation.}
}
\description{
Simple function to produce a \emph{FULL} dataset (all variables) for the \emph{observations}  from a model run in \code{lm()}. This is useful in running a nested regression.
}
\examples{

mod <- lm(hp ~ disp, data=mtcars)
modelData(mod)

}
