% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCorr.R
\name{pCorr}
\alias{pCorr}
\title{Partial and Semipartial (Part) Correlations}
\usage{
pCorr(obj)
}
\arguments{
\item{obj}{Model object from an \code{lm()} fitted equation.}
}
\description{
Simple function to produce partial and part (semipartial) correlation coefficients after a model run in \code{lm()}. Also returns the "squares" of each metric.
}
\note{
Requires at least \strong{two} independent variables in the \code{lm()} object.
}
\examples{
 mod <- lm(hp ~ disp + wt, data=mtcars)
 summary(mod)
 pCorr(mod)

}
