% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residFitted.R
\name{residFitted}
\alias{residFitted}
\title{Create a Residual vs. Fitted Plot}
\usage{
residFitted(obj, sigma = FALSE, id = FALSE, print = FALSE, ...)
}
\arguments{
\item{obj}{Model object form an lm() fitted equation.}

\item{sigma}{Level of residual Standard Deviation to plot. Default is none (FALSE).}

\item{id}{Logical: Should the identy() function be used to describe selected observations? Will automatically invoke print=TRUE for selected observations. Default is FALSE.}

\item{print}{Logical: Should observations outside the specified sigma level be printed to the console? Default is FALSE.}

\item{...}{Further arguments passed to and from other methods.}
}
\description{
Create a residual versus fitted values plot using base plot.
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
residFitted(mod, sigma=2, print=TRUE)
}
