% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDescribe.R
\name{simpleDescribe}
\alias{simpleDescribe}
\title{Simple Descriptive Statistics for Numerical Data}
\usage{
simpleDescribe(df, vars, rnd = 3)
}
\arguments{
\item{df}{Dataframe of variable location.}

\item{vars}{Variable of numerical nature.}

\item{rnd}{Rounding of output table (data.frame). Defaults to 3.}
}
\description{
Create a table of descriptive statistics for a \emph{single} numerical variable. This function is \code{dplyr} friendly (accepts piping).
}
\examples{
mtcars \%>\%
  group_by(cyl) \%>\%
  simpleDescribe(mpg)

}
