% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleScatter.R
\name{simpleScatter}
\alias{simpleScatter}
\title{Create a Simple ggplot Scatter Plot}
\usage{
simpleScatter(data, x, y, xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)), title = NULL, subtitle = NULL,
  xlim = NULL, ylim = NULL, line = FALSE, interval = "none", ...)
}
\arguments{
\item{data}{frame to be used.}

\item{x}{X (independent) variable to be plotted.}

\item{y}{Y (dependent) variable to be plotted.}

\item{xlab}{Label (optional) to be placed on the x axis.}

\item{ylab}{Label (optional) to be placed on the y axis.}

\item{title}{Title (optional) of the chart.}

\item{subtitle}{Subtitle (optional) of the chart.}

\item{xlim}{Typical xlim option to limit the graphed regions of the chart.}

\item{ylim}{Typical ylim option to limit the graphed regions of the chart.}

\item{line}{Logical: add linear fit line to plot?}

\item{interval}{Option denoting the type of interval to be plotted around linear prediction line. Option "none" is default. Other options include "confidence" and "prediction". Use of any option other than "none" will automatically change line=TRUE.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
ggplot object
}
\description{
A minimal function to produce a ggplot object from two continuous varaibles.
}
\examples{
m <- mtcars
simpleScatter(m, disp, hp, title="Horsepower from Displacement")
}

