% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studResidPlot.R
\name{studResidPlot}
\alias{studResidPlot}
\title{Create a Studentized Deleted Residual Plot}
\usage{
studResidPlot(obj, key.variable = NULL, print.obs = FALSE,
  print.plot = TRUE, sort.obs = TRUE, all.obs = FALSE)
}
\arguments{
\item{obj}{Object from an lm() fiitted equation.}

\item{key.variable}{Required if lm() data object is of tibble class. Name of the unique key variable (identifier variable). If data object is of data.frame class, this is optional and row.names will be used instead.}

\item{print.obs}{Logical: Should observations outside the specified sigma level be printed to the console?}

\item{print.plot}{Logical: Should plot be created?}

\item{sort.obs}{Logical: Should observations (if print.obs=TRUE) be sorted?}

\item{all.obs}{Logical: Should \strong{all} observations be printed?}
}
\description{
Produce a studentized deleted residual plot.
}
\examples{
mod <- lm(hp ~ disp, data=mtcars)
summary(mod)
studResidPlot(mod)
}
\seealso{
\code{\link{levPlot}}
\code{\link{cooksPlot}}
\code{\link{threeOuts}}
}
