% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagObs.R, R/tagObs_depreciated.R
\name{tagObs}
\alias{tagObs}
\title{Mark (tag) Observations in a Data Frame}
\usage{
tagObs(obj, tag = 1, name = NULL)

tagObs(obj, tag = 1, name = NULL)
}
\arguments{
\item{obj}{Model object generated by the lm() function.}

\item{tag}{Label with which to mark the used observations. Default is an indicator of 1.}

\item{name}{Optional name to assign to the suffix of the newly created varaible. Defualt is the model object name.}

\item{obj}{Model object generated by the lm() function.}
}
\description{
A minimal function to mark observations use in a prior run lm() model.

A minimal function to mark observations use in a prior run lm() model. Observations will be marked with a new variable in the original data used in the lm() function.
}
\examples{
mod_1 <- lm(hp ~ disp + wt, data=mtcars)
mtcars$In_mod1 <- tagObs(obj = mod_1)
mod_1 <- lm(hp ~ disp + wt, data=mtcars)
tagObs(obj = mod_1)
}

