% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagObs.R
\name{tagObs}
\alias{tagObs}
\title{Mark (tag) Observations in a Data Frame}
\usage{
tagObs(obj, tag = 1, name = NULL)
}
\arguments{
\item{obj}{Model object generated by the lm() function.}
}
\description{
A minimal function to mark observations use in a prior run lm() model.
}
\examples{
mod_1 <- lm(hp ~ disp + wt, data=mtcars)
mtcars$In_mod1 <- tagObs(obj = mod_1)
}

